requirements.txt
Flask
python-dotenv
requests
psycopg2-binary
SQLAlchemy
twitchio